Login

+++ gibz.m151.web/Controllers/PersonController.cs
@@ -1,5 +1,6 @@
 ﻿using gibz.m151.business;
 using gibz.m151.data.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -13,8 +14,14 @@ namespace WebApplication1.Controllers
        
         public IActionResult Index()
         {
-            List<Person> persons = Businesslogic.GetAllPersons();
-            return View(persons);
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                List<Person> persons = Businesslogic.GetAllPersons();
+                return View(persons);
+            }
+            {
+                return RedirectToAction("Login", "Login");
+            }
         }
         //[HttpGet]
         //public ActionResult GetPersons()
@@ -32,30 +39,59 @@ namespace WebApplication1.Controllers
         [HttpGet]
         public ActionResult GetPersonById()
         {
-            return View();
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Login", "Login");
+            }
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult GetPersonById(int id)
         {
-            Person person = gibz.m151.business.Businesslogic.GetPerson(id);
-            return View("Person", person);
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                Person person = gibz.m151.business.Businesslogic.GetPerson(id);
+                return View("Person", person);
+            }
+            else
+            {
+                return RedirectToAction("Login", "Login");
+            }
+
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Delete(Person person)
         {
-            Businesslogic.DeletePerson(person);
-            return RedirectToAction("Index");
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                Businesslogic.DeletePerson(person);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Login");
+            }
         }
 
         //shows the site with the values, if existing
         [HttpGet]
         public ActionResult Edit(int id)
         {
-            Person person = Businesslogic.ShowPersonToEdit(id);
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                Person person = Businesslogic.ShowPersonToEdit(id);
             return View(person);
+            }
+            else
+            {
+                return RedirectToAction("Login", "Login");
+            }
         }
 
 
@@ -64,8 +100,15 @@ namespace WebApplication1.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Person person)
         {
-            Businesslogic.EditPerson(person);
+            if (HttpContext.Session.GetString("UserID") != null)
+            {
+                Businesslogic.EditPerson(person);
             return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Login");
+            }
 
         }
